backend/requirements.txt
# The National Pulse – Backend
fastapi
uvicorn[standard]
feedparser
beautifulsoup4
requests
google-generativeai
python-dotenv
